+++ Samples~/Simple Character/SimpleCharacter.cs
@@ -6,6 +6,10 @@ using Lingotion.Thespeon.Inputs;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
+#if UNITY_EDITOR
+using Unity.Burst;
+#endif
+
 
 /// <summary>
 /// A simple character controller that uses the Thespeon engine for real-time voice synthesis.
@@ -22,6 +26,13 @@ public class SimpleCharacter : MonoBehaviour
     public ThespeonCharacterAsset actorAsset;
     void Start()
     {
+#if UNITY_EDITOR
+        if (BurstCompiler.Options.EnableBurstDebug)
+        {
+            Debug.LogWarning("[Warning] Burst Native Debug Mode Compilation is ON; performance will be slower in Editor when running Thespeon on CPU.");
+        }
+#endif
+
         engine = GetComponent<ThespeonEngine>();
         // Connect callback when audio is received from Thespeon
         engine.OnAudioReceived += OnAudioPacketReceive;