+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+# [1.3.0] - 2025-10-03
+## Added
+* License verification step to activate package use.
+## Fixed
+* Users are now reminded to turn of Burst Native Debug Mode Compilation when running Thespeon as it heavily affects model performance when run on CPU.
 # [1.2.0] - 2025-09-22
 ## Added
 * New `OnSynthesisFailed` callback for ThespeonEngine.